// du_website/src/Models/dashboardData.ts
type dashboardData = {
  stats: {
    total_outstanding: string;
    past_due: string;
    ytd_sales: string;
    last_ytd_sales: string;
  };
  recentOrders: {
    sales_number: string;
    order_date: string;
    transaction_status: number;
    transaction_header_id: string;
  }[];

  nextDeliveryDues: {
    delivery_number: string;
    due_date: string;
    transaction_status: number;
  }[];

  collectionHeaders: {
    invoice_number: string;
    due_date: string;
    amount: string;
  }[];

  openInvoice: {
    payment_amount: string;
    paid_for: string;
    payment_on: string;
    payment_type: string;
  };
  transaction_header: {
    order_amount: string;
    order_code: string;
    placed_on: string;
    sales_rep: string;
    next_planned_visit: string;
  };
  productsByCategory: {
    category: string;
    val: string;
  }[];
  productSales: {
    item: string;
    sold_quantity: number;
    variation: number;
    image_url: string;
  }[];

  seasonalVariation: {
    month: string;
    TotalSales: number;
    TotalReturns: number;
  }[];
};